Keep the campaign constants out of the type files
The delivery state, skip reason and failure reason each shipped their
constant and their derived type in one .type.ts file. Constants belong in
a .constant.ts of their own, the way the rest of the codebase does it, so
each file has one export and importing the type does not drag the value
along with it.
