+++ packages/twenty-server/src/engine/core-modules/emailing-domain/constants/campaign-delivery-state.constant.ts
@@ -1,0 +1,8 @@
+export const CAMPAIGN_DELIVERY_STATE = {
+  QUEUED: 'QUEUED',
+  SENDING: 'SENDING',
+  SENT: 'SENT',
+  FAILED: 'FAILED',
+  SKIPPED: 'SKIPPED',
+  UNCERTAIN: 'UNCERTAIN',
+} as const;
