+++ packages/twenty-server/src/engine/core-modules/emailing-domain/constants/campaign-skip-reason.constant.ts
@@ -1,0 +1,7 @@
+export const CAMPAIGN_SKIP_REASON = {
+  SUPPRESSED: 'SUPPRESSED',
+  OUT_OF_CREDITS: 'OUT_OF_CREDITS',
+  CAMPAIGN_CANCELED: 'CAMPAIGN_CANCELED',
+  REPUTATION_BLOCKED: 'REPUTATION_BLOCKED',
+  NO_EMAIL_ADDRESS: 'NO_EMAIL_ADDRESS',
+} as const;
