+++ packages/twenty-server/src/engine/core-modules/emailing-domain/types/campaign-delivery-state.type.ts
@@ -1,11 +1,4 @@
-export const CAMPAIGN_DELIVERY_STATE = {
-  QUEUED: 'QUEUED',
-  SENDING: 'SENDING',
-  SENT: 'SENT',
-  FAILED: 'FAILED',
-  SKIPPED: 'SKIPPED',
-  UNCERTAIN: 'UNCERTAIN',
-} as const;
+import { type CAMPAIGN_DELIVERY_STATE } from 'src/engine/core-modules/emailing-domain/constants/campaign-delivery-state.constant';
 
 export type CampaignDeliveryState =
   (typeof CAMPAIGN_DELIVERY_STATE)[keyof typeof CAMPAIGN_DELIVERY_STATE];
