+++ packages/twenty-server/src/engine/core-modules/emailing-domain/types/campaign-failure-reason.type.ts
@@ -1,17 +1,4 @@
-export const CAMPAIGN_FAILURE_REASON = {
-  TEMPORARY_ERROR: 'TEMPORARY_ERROR',
-  CONFIGURATION_ERROR: 'CONFIGURATION_ERROR',
-  INSUFFICIENT_PERMISSIONS: 'INSUFFICIENT_PERMISSIONS',
-  SANDBOX_ACCOUNT: 'SANDBOX_ACCOUNT',
-  SENDING_SUSPENDED: 'SENDING_SUSPENDED',
-  UNSUBSCRIBE_NOT_READY: 'UNSUBSCRIBE_NOT_READY',
-  UNSUBSCRIBE_MULTIPLE_RECIPIENTS: 'UNSUBSCRIBE_MULTIPLE_RECIPIENTS',
-  REJECTED: 'REJECTED',
-  RENDERING_FAILED: 'RENDERING_FAILED',
-  RETRY_EXHAUSTED: 'RETRY_EXHAUSTED',
-  CLAIM_EXPIRED: 'CLAIM_EXPIRED',
-  UNKNOWN: 'UNKNOWN',
-} as const;
+import { type CAMPAIGN_FAILURE_REASON } from 'src/engine/core-modules/emailing-domain/constants/campaign-failure-reason.constant';
 
 export type CampaignFailureReason =
   (typeof CAMPAIGN_FAILURE_REASON)[keyof typeof CAMPAIGN_FAILURE_REASON];
