+++ packages/twenty-server/src/engine/core-modules/emailing-domain/utils/__tests__/compute-campaign-counts.util.spec.ts
@@ -1,4 +1,4 @@
-import { CAMPAIGN_DELIVERY_STATE } from 'src/engine/core-modules/emailing-domain/types/campaign-delivery-state.type';
+import { CAMPAIGN_DELIVERY_STATE } from 'src/engine/core-modules/emailing-domain/constants/campaign-delivery-state.constant';
 import { computeCampaignCounts } from 'src/engine/core-modules/emailing-domain/utils/compute-campaign-counts.util';
 
 describe('computeCampaignCounts', () => {
